parse blob threshold to long, shows better error later in the code

+++ kvd-server/src/main/java/kvd/server/Kvd.java
@@ -135,7 +135,7 @@ public class Kvd {
         HumanReadable.parseDurationToMillisOrNull(options.expireAfterAccess, TimeUnit.SECONDS),
         HumanReadable.parseDurationToMillisOrNull(options.expireAfterWrite, TimeUnit.SECONDS),
         HumanReadable.parseDurationToMillisOrNull(options.expireCheckInterval, TimeUnit.SECONDS),
-        HumanReadableBytes.parseInt(options.blobThreshold),
+        HumanReadableBytes.parseLong(options.blobThreshold),
         HumanReadableBytes.parseLong(options.blobSplitThreshold));
   }
